+++ src/main/java/com/arangodb/springframework/core/convert/DBDocumentEntity.java
@@ -27,6 +27,7 @@ import java.util.Map;
  * @author Mark Vollmary
  *
  */
+@SuppressWarnings("deprecation")
 public class DBDocumentEntity extends HashMap<String, Object> implements DBEntity {
 
 	private static final long serialVersionUID = -7251842887063588024L;
@@ -39,7 +40,11 @@ public class DBDocumentEntity extends HashMap<String, Object> implements DBEntit
 		super(m);
 	}
 
+	/**
+	 * @deprecated Will be removed in 3.0.0
+	 */
 	@Override
+	@Deprecated
 	public boolean add(final Object value) {
 		throw new UnsupportedOperationException();
 	}