+++ src/main/java/com/arangodb/springframework/core/template/ArangoTemplate.java
@@ -81,6 +81,7 @@ import com.arangodb.util.MapBuilder;
  * @author Christian Lechner
  *
  */
+@SuppressWarnings("deprecation")
 public class ArangoTemplate implements ArangoOperations, CollectionCallback {
 
 	private volatile ArangoDBVersion version;
@@ -100,10 +101,10 @@ public class ArangoTemplate implements ArangoOperations, CollectionCallback {
 	}
 
 	public ArangoTemplate(final ArangoDB.Builder arango, final String database, final ArangoConverter converter,
-			final PersistenceExceptionTranslator exceptionTranslator) {
+		final PersistenceExceptionTranslator exceptionTranslator) {
 		super();
-		this.arango = arango.build()
-				._setCursorInitializer(new com.arangodb.springframework.core.template.ArangoCursorInitializer(converter));
+		this.arango = arango.build()._setCursorInitializer(
+			new com.arangodb.springframework.core.template.ArangoCursorInitializer(converter));
 		this.databaseName = database;
 		this.converter = converter;
 		this.exceptionTranslator = exceptionTranslator;
@@ -157,13 +158,16 @@ public class ArangoTemplate implements ArangoOperations, CollectionCallback {
 	}
 
 	private ArangoCollection _collection(final Class<?> entityClass, final String id) {
-		final ArangoPersistentEntity<?> persistentEntity = converter.getMappingContext().getPersistentEntity(entityClass);
+		final ArangoPersistentEntity<?> persistentEntity = converter.getMappingContext()
+				.getPersistentEntity(entityClass);
 		final String name = determineCollectionFromId(Optional.ofNullable(id)).orElse(persistentEntity.getCollection());
 		return _collection(name, persistentEntity, persistentEntity.getCollectionOptions());
 	}
 
-	private ArangoCollection _collection(final String name, final ArangoPersistentEntity<?> persistentEntity,
-			final CollectionCreateOptions options) {
+	private ArangoCollection _collection(
+		final String name,
+		final ArangoPersistentEntity<?> persistentEntity,
+		final CollectionCreateOptions options) {
 
 		return collectionCache.computeIfAbsent(name, collName -> {
 			final ArangoCollection collection = db().collection(collName);
@@ -187,8 +191,9 @@ public class ArangoTemplate implements ArangoOperations, CollectionCallback {
 		});
 	}
 
-	private static void ensureCollectionIndexes(final CollectionOperations collection,
-			final ArangoPersistentEntity<?> persistentEntity) {
+	private static void ensureCollectionIndexes(
+		final CollectionOperations collection,
+		final ArangoPersistentEntity<?> persistentEntity) {
 		persistentEntity.getHashIndexes().stream().forEach(index -> ensureHashIndex(collection, index));
 		persistentEntity.getHashIndexedProperties().stream().forEach(p -> ensureHashIndex(collection, p));
 		persistentEntity.getSkiplistIndexes().stream().forEach(index -> ensureSkiplistIndex(collection, index));
@@ -202,13 +207,14 @@ public class ArangoTemplate implements ArangoOperations, CollectionCallback {
 	}
 
 	private static void ensureHashIndex(final CollectionOperations collection, final HashIndex annotation) {
-		collection.ensureHashIndex(Arrays.asList(annotation.fields()), new HashIndexOptions().unique(annotation.unique())
-				.sparse(annotation.sparse()).deduplicate(annotation.deduplicate()));
+		collection.ensureHashIndex(Arrays.asList(annotation.fields()), new HashIndexOptions()
+				.unique(annotation.unique()).sparse(annotation.sparse()).deduplicate(annotation.deduplicate()));
 	}
 
 	private static void ensureHashIndex(final CollectionOperations collection, final ArangoPersistentProperty value) {
 		final HashIndexOptions options = new HashIndexOptions();
-		value.getHashIndexed().ifPresent(i -> options.unique(i.unique()).sparse(i.sparse()).deduplicate(i.deduplicate()));
+		value.getHashIndexed()
+				.ifPresent(i -> options.unique(i.unique()).sparse(i.sparse()).deduplicate(i.deduplicate()));
 		collection.ensureHashIndex(Collections.singleton(value.getFieldName()), options);
 	}
 
@@ -217,7 +223,9 @@ public class ArangoTemplate implements ArangoOperations, CollectionCallback {
 				.unique(annotation.unique()).sparse(annotation.sparse()).deduplicate(annotation.deduplicate()));
 	}
 
-	private static void ensureSkiplistIndex(final CollectionOperations collection, final ArangoPersistentProperty value) {
+	private static void ensureSkiplistIndex(
+		final CollectionOperations collection,
+		final ArangoPersistentProperty value) {
 		final SkiplistIndexOptions options = new SkiplistIndexOptions();
 		value.getSkiplistIndexed()
 				.ifPresent(i -> options.unique(i.unique()).sparse(i.sparse()).deduplicate(i.deduplicate()));
@@ -226,18 +234,20 @@ public class ArangoTemplate implements ArangoOperations, CollectionCallback {
 
 	private static void ensurePersistentIndex(final CollectionOperations collection, final PersistentIndex annotation) {
 		collection.ensurePersistentIndex(Arrays.asList(annotation.fields()),
-				new PersistentIndexOptions().unique(annotation.unique()).sparse(annotation.sparse()));
+			new PersistentIndexOptions().unique(annotation.unique()).sparse(annotation.sparse()));
 	}
 
-	private static void ensurePersistentIndex(final CollectionOperations collection,
-			final ArangoPersistentProperty value) {
+	private static void ensurePersistentIndex(
+		final CollectionOperations collection,
+		final ArangoPersistentProperty value) {
 		final PersistentIndexOptions options = new PersistentIndexOptions();
 		value.getPersistentIndexed().ifPresent(i -> options.unique(i.unique()).sparse(i.sparse()));
 		collection.ensurePersistentIndex(Collections.singleton(value.getFieldName()), options);
 	}
 
 	private static void ensureGeoIndex(final CollectionOperations collection, final GeoIndex annotation) {
-		collection.ensureGeoIndex(Arrays.asList(annotation.fields()), new GeoIndexOptions().geoJson(annotation.geoJson()));
+		collection.ensureGeoIndex(Arrays.asList(annotation.fields()),
+			new GeoIndexOptions().geoJson(annotation.geoJson()));
 	}
 
 	private static void ensureGeoIndex(final CollectionOperations collection, final ArangoPersistentProperty value) {
@@ -248,10 +258,12 @@ public class ArangoTemplate implements ArangoOperations, CollectionCallback {
 
 	private static void ensureFulltextIndex(final CollectionOperations collection, final FulltextIndex annotation) {
 		collection.ensureFulltextIndex(Collections.singleton(annotation.field()),
-				new FulltextIndexOptions().minLength(annotation.minLength() > -1 ? annotation.minLength() : null));
+			new FulltextIndexOptions().minLength(annotation.minLength() > -1 ? annotation.minLength() : null));
 	}
 
-	private static void ensureFulltextIndex(final CollectionOperations collection, final ArangoPersistentProperty value) {
+	private static void ensureFulltextIndex(
+		final CollectionOperations collection,
+		final ArangoPersistentProperty value) {
 		final FulltextIndexOptions options = new FulltextIndexOptions();
 		value.getFulltextIndexed().ifPresent(i -> options.minLength(i.minLength() > -1 ? i.minLength() : null));
 		collection.ensureFulltextIndex(Collections.singleton(value.getFieldName()), options);
@@ -304,11 +316,14 @@ public class ArangoTemplate implements ArangoOperations, CollectionCallback {
 	}
 
 	@Override
-	public <T> ArangoCursor<T> query(final String query, final Map<String, Object> bindVars,
-			final AqlQueryOptions options, final Class<T> entityClass) throws DataAccessException {
+	public <T> ArangoCursor<T> query(
+		final String query,
+		final Map<String, Object> bindVars,
+		final AqlQueryOptions options,
+		final Class<T> entityClass) throws DataAccessException {
 		return db().query(query,
-				bindVars == null ? null : DBDocumentEntity.class.cast(toDBEntity(prepareBindVars(bindVars))), options,
-				entityClass);
+			bindVars == null ? null : DBDocumentEntity.class.cast(toDBEntity(prepareBindVars(bindVars))), options,
+			entityClass);
 	}
 
 	private Map<String, Object> prepareBindVars(final Map<String, Object> bindVars) {
@@ -321,19 +336,22 @@ public class ArangoTemplate implements ArangoOperations, CollectionCallback {
 	}
 
 	@Override
-	public MultiDocumentEntity<? extends DocumentEntity> delete(final Iterable<Object> values, final Class<?> entityClass,
-			final DocumentDeleteOptions options) throws DataAccessException {
+	public MultiDocumentEntity<? extends DocumentEntity> delete(
+		final Iterable<Object> values,
+		final Class<?> entityClass,
+		final DocumentDeleteOptions options) throws DataAccessException {
 		try {
-			return _collection(entityClass).deleteDocuments(DBCollectionEntity.class.cast(toDBEntity(values)), entityClass,
-					options);
+			return _collection(entityClass).deleteDocuments(DBCollectionEntity.class.cast(toDBEntity(values)),
+				entityClass, options);
 		} catch (final ArangoDBException e) {
 			throw translateExceptionIfPossible(e);
 		}
 	}
 
 	@Override
-	public MultiDocumentEntity<? extends DocumentEntity> delete(final Iterable<Object> values, final Class<?> entityClass)
-			throws DataAccessException {
+	public MultiDocumentEntity<? extends DocumentEntity> delete(
+		final Iterable<Object> values,
+		final Class<?> entityClass) throws DataAccessException {
 		return delete(values, entityClass, new DocumentDeleteOptions());
 	}
 
@@ -353,8 +371,10 @@ public class ArangoTemplate implements ArangoOperations, CollectionCallback {
 	}
 
 	@Override
-	public <T> MultiDocumentEntity<? extends DocumentEntity> update(final Iterable<T> values, final Class<T> entityClass,
-			final DocumentUpdateOptions options) throws DataAccessException {
+	public <T> MultiDocumentEntity<? extends DocumentEntity> update(
+		final Iterable<T> values,
+		final Class<T> entityClass,
+		final DocumentUpdateOptions options) throws DataAccessException {
 		try {
 			final MultiDocumentEntity<? extends DocumentEntity> res = _collection(entityClass)
 					.updateDocuments(DBCollectionEntity.class.cast(toDBEntity(values)), options);
@@ -366,8 +386,9 @@ public class ArangoTemplate implements ArangoOperations, CollectionCallback {
 	}
 
 	@Override
-	public <T> MultiDocumentEntity<? extends DocumentEntity> update(final Iterable<T> values, final Class<T> entityClass)
-			throws DataAccessException {
+	public <T> MultiDocumentEntity<? extends DocumentEntity> update(
+		final Iterable<T> values,
+		final Class<T> entityClass) throws DataAccessException {
 		return update(values, entityClass, new DocumentUpdateOptions());
 	}
 
@@ -376,7 +397,7 @@ public class ArangoTemplate implements ArangoOperations, CollectionCallback {
 			throws DataAccessException {
 		try {
 			final DocumentEntity res = _collection(value.getClass(), id).updateDocument(determineDocumentKeyFromId(id),
-					toDBEntity(value), options);
+				toDBEntity(value), options);
 			updateDBFields(value, res);
 			return res;
 		} catch (final ArangoDBException e) {
@@ -390,8 +411,10 @@ public class ArangoTemplate implements ArangoOperations, CollectionCallback {
 	}
 
 	@Override
-	public <T> MultiDocumentEntity<? extends DocumentEntity> replace(final Iterable<T> values, final Class<T> entityClass,
-			final DocumentReplaceOptions options) throws DataAccessException {
+	public <T> MultiDocumentEntity<? extends DocumentEntity> replace(
+		final Iterable<T> values,
+		final Class<T> entityClass,
+		final DocumentReplaceOptions options) throws DataAccessException {
 		try {
 			final MultiDocumentEntity<? extends DocumentEntity> res = _collection(entityClass)
 					.replaceDocuments(DBCollectionEntity.class.cast(toDBEntity(values)), options);
@@ -403,8 +426,9 @@ public class ArangoTemplate implements ArangoOperations, CollectionCallback {
 	}
 
 	@Override
-	public <T> MultiDocumentEntity<? extends DocumentEntity> replace(final Iterable<T> values, final Class<T> entityClass)
-			throws DataAccessException {
+	public <T> MultiDocumentEntity<? extends DocumentEntity> replace(
+		final Iterable<T> values,
+		final Class<T> entityClass) throws DataAccessException {
 		return replace(values, entityClass, new DocumentReplaceOptions());
 	}
 
@@ -413,7 +437,7 @@ public class ArangoTemplate implements ArangoOperations, CollectionCallback {
 			throws DataAccessException {
 		try {
 			final DocumentEntity res = _collection(value.getClass(), id).replaceDocument(determineDocumentKeyFromId(id),
-					toDBEntity(value), options);
+				toDBEntity(value), options);
 			updateDBFields(value, res);
 			return res;
 		} catch (final ArangoDBException e) {
@@ -430,8 +454,8 @@ public class ArangoTemplate implements ArangoOperations, CollectionCallback {
 	public <T> Optional<T> find(final String id, final Class<T> entityClass, final DocumentReadOptions options)
 			throws DataAccessException {
 		try {
-			final DBEntity doc = _collection(entityClass, id).getDocument(determineDocumentKeyFromId(id), DBEntity.class,
-					options);
+			final DBEntity doc = _collection(entityClass, id).getDocument(determineDocumentKeyFromId(id),
+				DBEntity.class, options);
 			return Optional.ofNullable(fromDBEntity(entityClass, doc));
 		} catch (final ArangoDBException e) {
 			throw translateExceptionIfPossible(e);
@@ -467,8 +491,10 @@ public class ArangoTemplate implements ArangoOperations, CollectionCallback {
 	}
 
 	@Override
-	public <T> MultiDocumentEntity<? extends DocumentEntity> insert(final Iterable<T> values, final Class<T> entityClass,
-			final DocumentCreateOptions options) throws DataAccessException {
+	public <T> MultiDocumentEntity<? extends DocumentEntity> insert(
+		final Iterable<T> values,
+		final Class<T> entityClass,
+		final DocumentCreateOptions options) throws DataAccessException {
 		try {
 			final MultiDocumentEntity<? extends DocumentEntity> res = _collection(entityClass)
 					.insertDocuments(DBCollectionEntity.class.cast(toDBEntity(values)), options);
@@ -480,8 +506,9 @@ public class ArangoTemplate implements ArangoOperations, CollectionCallback {
 	}
 
 	@Override
-	public <T> MultiDocumentEntity<? extends DocumentEntity> insert(final Iterable<T> values, final Class<T> entityClass)
-			throws DataAccessException {
+	public <T> MultiDocumentEntity<? extends DocumentEntity> insert(
+		final Iterable<T> values,
+		final Class<T> entityClass) throws DataAccessException {
 		return insert(values, entityClass, new DocumentCreateOptions());
 	}